+++ CHANGELOG.md
@@ -12,6 +12,7 @@
  *   New URL param step= and API call setExecutionStep() to enable/disable debug stepping and how many instructions for each step
  *   New URL param e= to allow base64 encoded params in it (circumvent mod_security conflicts)
  *   getX() and getY() anow return an integer to allow proper validation if "same coordinate" after moving around the guide
+ *   Custom instructions can now include the code in the instructions' definitions instead of delegating it to precode
  *   Updated create.js
  *   Fix: Maximizing console is back working, and better than every (full maximization)